+++ plone/app/contenttypes/tests/test_migration.py
@@ -8,12 +8,12 @@ from plone.app.contenttypes.testing import set_browserlayer
 from plone.event.interfaces import IEventAccessor
 from plone.app.testing import login
 from plone.app.testing import applyProfile
+from zope.annotation.interfaces import IAnnotations
 from zope.component import getMultiAdapter
 from zope.component import getSiteManager
 from zope.component import getUtility
 from zope.intid.interfaces import IIntIds
 from zope.schema.interfaces import IVocabularyFactory
-
 import os.path
 import time
 import unittest2 as unittest
@@ -122,7 +122,6 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         """Can we migrate a Products.ATContentTypes event?"""
         from DateTime import DateTime
         from plone.app.contenttypes.migration.migration import migrate_events
-        from plone.app.event.dx.behaviors import IEventSummary
 
         # create an ATEvent
         self.portal.invokeFactory('Event', 'event')
@@ -190,7 +189,7 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         self.assertEqual(('You', 'Me'), dx_acc.attendees)
         self.assertEquals('Event', dx_event.__class__.__name__)
         self.assertEqual(u'<p>T\xfctensuppe</p>', dx_acc.text)
-        self.assertEqual(u'Tütensuppe', IEventSummary(dx_event).text.raw)
+        self.assertEqual(u'Tütensuppe', dx_event.text.raw)
 
     def test_pae_atevent_is_migrated(self):
         """Can we migrate a plone.app.event AT event?"""
@@ -198,7 +197,6 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         from plone.testing import z2
         from plone.app.testing import applyProfile
         from plone.app.contenttypes.migration.migration import migrate_events
-        from plone.app.event.dx.behaviors import IEventSummary
 
         # Enable plone.app.event.at
         z2.installProduct(self.layer['app'], 'plone.app.event.at')
@@ -261,18 +259,17 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         self.assertEqual(u'dummy@email.com', new_event_acc.contact_email)
         self.assertEqual(u'123456789', new_event_acc.contact_phone)
         self.assertEqual(u'http://www.plone.org', new_event_acc.event_url)
-        self.assertEqual(u'<p>T\xfctensuppe</p>', new_event_acc.text)
-        self.assertEqual(u'Tütensuppe', IEventSummary(new_event).text.raw)
+        self.assertEqual(u'<p>T\xfctensuppe</p>', new_event.text.output)
+        self.assertEqual(u'Tütensuppe', new_event.text.raw)
 
     def test_pae_dxevent_is_migrated(self):
         from datetime import datetime
         from plone.app.contenttypes.migration.migration import migrate_events
         from plone.app.textfield.value import RichTextValue
-        from plone.app.event.dx.behaviors import IEventSummary
 
         # Enable plone.app.event.dx
         from plone.app.testing import applyProfile
-        applyProfile(self.portal, 'plone.app.event.dx:default')
+        applyProfile(self.portal, 'plone.app.event:testing')
 
         old_event = self.portal[self.portal.invokeFactory(
             'plone.app.event.dx.event',
@@ -287,11 +284,15 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         old_event_acc.contact_email = 'me@geor.ge'
         old_event_acc.contact_phone = '+99512345'
         old_event_acc.event_url = 'http://geor.ge/event'
-        old_event_acc.text = RichTextValue(
+        # We need to manually place the value of the "text" field into
+        # annotation storage
+        richtext = RichTextValue(
             raw='Woo, yeah',
             mimeType='text/plain',
             outputMimeType='text/x-html-safe'
         )
+        ann = IAnnotations(old_event)
+        ann['plone.app.event.dx.behaviors.IEventSummary.text'] = richtext
 
         # migrate
         applyProfile(self.portal, 'plone.app.contenttypes:default')
@@ -317,8 +318,8 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         self.assertEqual(u'me@geor.ge', new_event_acc.contact_email)
         self.assertEqual(u'+99512345', new_event_acc.contact_phone)
         self.assertEqual(u'http://geor.ge/event', new_event_acc.event_url)
-        self.assertEqual(u'<p>Woo, yeah</p>', new_event_acc.text)
-        self.assertEqual('Woo, yeah', IEventSummary(new_event).text.raw)
+        self.assertEqual(u'<p>Woo, yeah</p>', new_event.text.output)
+        self.assertEqual('Woo, yeah', new_event.text.raw)
         self.assertEqual(False, new_event.exclude_from_nav)
 
     def test_pact_1_0_dxevent_is_migrated(self):
@@ -326,7 +327,6 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         from pytz import timezone
         from plone.app.contenttypes.migration.migration import migrate_events
         from plone.app.textfield.value import RichTextValue
-        from plone.app.event.dx.behaviors import IEventSummary
         from plone.app.contenttypes.tests.oldtypes import create1_0EventType
 
         # Create a 1.0 Event object
@@ -376,8 +376,8 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         self.assertEqual(u'me@fra.nk', new_event_acc.contact_email)
         self.assertEqual(u'+4412345', new_event_acc.contact_phone)
         self.assertEqual(u'http://woo.com', new_event_acc.event_url)
-        self.assertEqual(u'<p>Awesüme</p>', new_event_acc.text)
-        self.assertEqual(u'Awesüme', IEventSummary(new_event).text.raw)
+        self.assertEqual(u'<p>Awesüme</p>', new_event.text.output)
+        self.assertEqual(u'Awesüme', new_event.text.raw)
         self.assertEqual(False, new_event.exclude_from_nav)
 
     def test_dx_excl_from_nav_is_migrated(self):
@@ -386,7 +386,7 @@ class MigrateToATContentTypesTest(unittest.TestCase):
 
         # Enable plone.app.event.dx
         from plone.app.testing import applyProfile
-        applyProfile(self.portal, 'plone.app.event.dx:default')
+        applyProfile(self.portal, 'plone.app.event:testing')
 
         old_event = self.portal[self.portal.invokeFactory(
             'plone.app.event.dx.event',
@@ -462,18 +462,28 @@ class MigrateToATContentTypesTest(unittest.TestCase):
         self.assertTrue(at_document is not dx_document)
 
     def test_collection_is_migrated(self):
-        from plone.app.contenttypes.migration.migration import\
-            CollectionMigrator
+        from plone.app.contenttypes.migration.migration import \
+            migrate_collections
+        from plone.app.contenttypes.behaviors.collection import \
+            ICollection as ICollectionBehavior
         from plone.app.contenttypes.interfaces import ICollection
-        if 'Collection' in self.portal.portal_types.keys():
-            self.portal.invokeFactory('Collection', 'collection')
-            at_collection = self.portal['collection']
-            applyProfile(self.portal, 'plone.app.contenttypes:default')
-            migrator = self.get_migrator(at_collection, CollectionMigrator)
-            migrator.migrate()
-            dx_collection = self.portal['collection']
-            self.assertTrue(ICollection.providedBy(dx_collection))
-            self.assertTrue(at_collection is not dx_collection)
+        self.portal.invokeFactory('Collection', 'collection')
+        at_collection = self.portal['collection']
+        at_collection.setText("<p>Whopee</p>")
+        query = [{
+            'i': 'Type',
+            'o': 'plone.app.querystring.operation.string.is',
+            'v': 'Document',
+        }]
+        at_collection.setQuery(query)
+        applyProfile(self.portal, 'plone.app.contenttypes:default')
+        migrate_collections(self.portal)
+        dx_collection = self.portal['collection']
+        self.assertTrue(ICollection.providedBy(dx_collection))
+        self.assertTrue(at_collection is not dx_collection)
+        wrapped = ICollectionBehavior(dx_collection)
+        self.assertEqual(wrapped.query, query)
+        self.assertEqual(dx_collection.text.output, "<p>Whopee</p>")
 
     def test_document_content_is_migrated(self):
         from plone.app.contenttypes.migration.migration import DocumentMigrator